@@ -1,10 +1,9 @@
 name: google_maps_flutter_web_integration_tests
 publish_to: none
 
-# Tests require flutter beta or greater to run.
 environment:
-  sdk: ^3.3.0
-  flutter: ">=3.19.0"
+  sdk: ^3.4.0
+  flutter: ">=3.22.0"
 
 dependencies:
   flutter:
@@ -12,15 +11,15 @@ dependencies:
   google_maps_flutter_platform_interface: ^2.7.0
   google_maps_flutter_web:
     path: ../
-  web: ^0.5.0
+  web: ^1.0.0
 
 dev_dependencies:
   build_runner: ^2.1.1
   flutter_test:
     sdk: flutter
-  google_maps: ^7.1.0
+  google_maps: ^8.0.0
   google_maps_flutter: ^2.2.0 # Needed for projection_test.dart
-  http: ">=0.13.0 <2.0.0"
+  http: ^1.2.2
   integration_test:
     sdk: flutter
   mockito: 5.4.4
