@@ -3,13 +3,15 @@
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i5;
+import 'dart:async' as _i6;
 
-import 'package:google_maps/google_maps.dart' as _i2;
+import 'package:google_maps/google_maps.dart' as _i4;
 import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart'
-    as _i4;
-import 'package:google_maps_flutter_web/google_maps_flutter_web.dart' as _i3;
+    as _i3;
+import 'package:google_maps_flutter_web/google_maps_flutter_web.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
+
+import 'google_maps_controller_test.dart' as _i5;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
@@ -24,52 +26,40 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeGMap_0 extends _i1.SmartFake implements _i2.GMap {
-  _FakeGMap_0(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
 /// A class which mocks [CirclesController].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockCirclesController extends _i1.Mock implements _i3.CirclesController {
+class MockCirclesController extends _i1.Mock implements _i2.CirclesController {
   @override
-  Map<_i4.CircleId, _i3.CircleController> get circles => (super.noSuchMethod(
+  Map<_i3.CircleId, _i2.CircleController> get circles => (super.noSuchMethod(
         Invocation.getter(#circles),
-        returnValue: <_i4.CircleId, _i3.CircleController>{},
-        returnValueForMissingStub: <_i4.CircleId, _i3.CircleController>{},
-      ) as Map<_i4.CircleId, _i3.CircleController>);
+        returnValue: <_i3.CircleId, _i2.CircleController>{},
+        returnValueForMissingStub: <_i3.CircleId, _i2.CircleController>{},
+      ) as Map<_i3.CircleId, _i2.CircleController>);
+
   @override
-  _i2.GMap get googleMap => (super.noSuchMethod(
+  _i4.Map get googleMap => (super.noSuchMethod(
         Invocation.getter(#googleMap),
-        returnValue: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-        returnValueForMissingStub: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-      ) as _i2.GMap);
+        returnValue: _i5.mapShim(),
+        returnValueForMissingStub: _i5.mapShim(),
+      ) as _i4.Map);
+
   @override
-  set googleMap(_i2.GMap? _googleMap) => super.noSuchMethod(
+  set googleMap(_i4.Map? _googleMap) => super.noSuchMethod(
         Invocation.setter(
           #googleMap,
           _googleMap,
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get mapId => (super.noSuchMethod(
         Invocation.getter(#mapId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
+
   @override
   set mapId(int? _mapId) => super.noSuchMethod(
         Invocation.setter(
@@ -78,24 +68,27 @@ class MockCirclesController extends _i1.Mock implements _i3.CirclesController {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void addCircles(Set<_i4.Circle>? circlesToAdd) => super.noSuchMethod(
+  void addCircles(Set<_i3.Circle>? circlesToAdd) => super.noSuchMethod(
         Invocation.method(
           #addCircles,
           [circlesToAdd],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void changeCircles(Set<_i4.Circle>? circlesToChange) => super.noSuchMethod(
+  void changeCircles(Set<_i3.Circle>? circlesToChange) => super.noSuchMethod(
         Invocation.method(
           #changeCircles,
           [circlesToChange],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void removeCircles(Set<_i4.CircleId>? circleIdsToRemove) =>
+  void removeCircles(Set<_i3.CircleId>? circleIdsToRemove) =>
       super.noSuchMethod(
         Invocation.method(
           #removeCircles,
@@ -103,10 +96,11 @@ class MockCirclesController extends _i1.Mock implements _i3.CirclesController {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void bindToMap(
     int? mapId,
-    _i2.GMap? googleMap,
+    _i4.Map? googleMap,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -124,39 +118,37 @@ class MockCirclesController extends _i1.Mock implements _i3.CirclesController {
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockPolygonsController extends _i1.Mock
-    implements _i3.PolygonsController {
+    implements _i2.PolygonsController {
   @override
-  Map<_i4.PolygonId, _i3.PolygonController> get polygons => (super.noSuchMethod(
+  Map<_i3.PolygonId, _i2.PolygonController> get polygons => (super.noSuchMethod(
         Invocation.getter(#polygons),
-        returnValue: <_i4.PolygonId, _i3.PolygonController>{},
-        returnValueForMissingStub: <_i4.PolygonId, _i3.PolygonController>{},
-      ) as Map<_i4.PolygonId, _i3.PolygonController>);
+        returnValue: <_i3.PolygonId, _i2.PolygonController>{},
+        returnValueForMissingStub: <_i3.PolygonId, _i2.PolygonController>{},
+      ) as Map<_i3.PolygonId, _i2.PolygonController>);
+
   @override
-  _i2.GMap get googleMap => (super.noSuchMethod(
+  _i4.Map get googleMap => (super.noSuchMethod(
         Invocation.getter(#googleMap),
-        returnValue: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-        returnValueForMissingStub: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-      ) as _i2.GMap);
+        returnValue: _i5.mapShim(),
+        returnValueForMissingStub: _i5.mapShim(),
+      ) as _i4.Map);
+
   @override
-  set googleMap(_i2.GMap? _googleMap) => super.noSuchMethod(
+  set googleMap(_i4.Map? _googleMap) => super.noSuchMethod(
         Invocation.setter(
           #googleMap,
           _googleMap,
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get mapId => (super.noSuchMethod(
         Invocation.getter(#mapId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
+
   @override
   set mapId(int? _mapId) => super.noSuchMethod(
         Invocation.setter(
@@ -165,24 +157,27 @@ class MockPolygonsController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void addPolygons(Set<_i4.Polygon>? polygonsToAdd) => super.noSuchMethod(
+  void addPolygons(Set<_i3.Polygon>? polygonsToAdd) => super.noSuchMethod(
         Invocation.method(
           #addPolygons,
           [polygonsToAdd],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void changePolygons(Set<_i4.Polygon>? polygonsToChange) => super.noSuchMethod(
+  void changePolygons(Set<_i3.Polygon>? polygonsToChange) => super.noSuchMethod(
         Invocation.method(
           #changePolygons,
           [polygonsToChange],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void removePolygons(Set<_i4.PolygonId>? polygonIdsToRemove) =>
+  void removePolygons(Set<_i3.PolygonId>? polygonIdsToRemove) =>
       super.noSuchMethod(
         Invocation.method(
           #removePolygons,
@@ -190,10 +185,11 @@ class MockPolygonsController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void bindToMap(
     int? mapId,
-    _i2.GMap? googleMap,
+    _i4.Map? googleMap,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -211,39 +207,37 @@ class MockPolygonsController extends _i1.Mock
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockPolylinesController extends _i1.Mock
-    implements _i3.PolylinesController {
+    implements _i2.PolylinesController {
   @override
-  Map<_i4.PolylineId, _i3.PolylineController> get lines => (super.noSuchMethod(
+  Map<_i3.PolylineId, _i2.PolylineController> get lines => (super.noSuchMethod(
         Invocation.getter(#lines),
-        returnValue: <_i4.PolylineId, _i3.PolylineController>{},
-        returnValueForMissingStub: <_i4.PolylineId, _i3.PolylineController>{},
-      ) as Map<_i4.PolylineId, _i3.PolylineController>);
+        returnValue: <_i3.PolylineId, _i2.PolylineController>{},
+        returnValueForMissingStub: <_i3.PolylineId, _i2.PolylineController>{},
+      ) as Map<_i3.PolylineId, _i2.PolylineController>);
+
   @override
-  _i2.GMap get googleMap => (super.noSuchMethod(
+  _i4.Map get googleMap => (super.noSuchMethod(
         Invocation.getter(#googleMap),
-        returnValue: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-        returnValueForMissingStub: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-      ) as _i2.GMap);
+        returnValue: _i5.mapShim(),
+        returnValueForMissingStub: _i5.mapShim(),
+      ) as _i4.Map);
+
   @override
-  set googleMap(_i2.GMap? _googleMap) => super.noSuchMethod(
+  set googleMap(_i4.Map? _googleMap) => super.noSuchMethod(
         Invocation.setter(
           #googleMap,
           _googleMap,
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get mapId => (super.noSuchMethod(
         Invocation.getter(#mapId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
+
   @override
   set mapId(int? _mapId) => super.noSuchMethod(
         Invocation.setter(
@@ -252,16 +246,18 @@ class MockPolylinesController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void addPolylines(Set<_i4.Polyline>? polylinesToAdd) => super.noSuchMethod(
+  void addPolylines(Set<_i3.Polyline>? polylinesToAdd) => super.noSuchMethod(
         Invocation.method(
           #addPolylines,
           [polylinesToAdd],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void changePolylines(Set<_i4.Polyline>? polylinesToChange) =>
+  void changePolylines(Set<_i3.Polyline>? polylinesToChange) =>
       super.noSuchMethod(
         Invocation.method(
           #changePolylines,
@@ -269,8 +265,9 @@ class MockPolylinesController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void removePolylines(Set<_i4.PolylineId>? polylineIdsToRemove) =>
+  void removePolylines(Set<_i3.PolylineId>? polylineIdsToRemove) =>
       super.noSuchMethod(
         Invocation.method(
           #removePolylines,
@@ -278,10 +275,11 @@ class MockPolylinesController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void bindToMap(
     int? mapId,
-    _i2.GMap? googleMap,
+    _i4.Map? googleMap,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -298,39 +296,37 @@ class MockPolylinesController extends _i1.Mock
 /// A class which mocks [MarkersController].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockMarkersController extends _i1.Mock implements _i3.MarkersController {
+class MockMarkersController extends _i1.Mock implements _i2.MarkersController {
   @override
-  Map<_i4.MarkerId, _i3.MarkerController> get markers => (super.noSuchMethod(
+  Map<_i3.MarkerId, _i2.MarkerController> get markers => (super.noSuchMethod(
         Invocation.getter(#markers),
-        returnValue: <_i4.MarkerId, _i3.MarkerController>{},
-        returnValueForMissingStub: <_i4.MarkerId, _i3.MarkerController>{},
-      ) as Map<_i4.MarkerId, _i3.MarkerController>);
+        returnValue: <_i3.MarkerId, _i2.MarkerController>{},
+        returnValueForMissingStub: <_i3.MarkerId, _i2.MarkerController>{},
+      ) as Map<_i3.MarkerId, _i2.MarkerController>);
+
   @override
-  _i2.GMap get googleMap => (super.noSuchMethod(
+  _i4.Map get googleMap => (super.noSuchMethod(
         Invocation.getter(#googleMap),
-        returnValue: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-        returnValueForMissingStub: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-      ) as _i2.GMap);
+        returnValue: _i5.mapShim(),
+        returnValueForMissingStub: _i5.mapShim(),
+      ) as _i4.Map);
+
   @override
-  set googleMap(_i2.GMap? _googleMap) => super.noSuchMethod(
+  set googleMap(_i4.Map? _googleMap) => super.noSuchMethod(
         Invocation.setter(
           #googleMap,
           _googleMap,
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get mapId => (super.noSuchMethod(
         Invocation.getter(#mapId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
+
   @override
   set mapId(int? _mapId) => super.noSuchMethod(
         Invocation.setter(
@@ -339,28 +335,31 @@ class MockMarkersController extends _i1.Mock implements _i3.MarkersController {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  _i5.Future<void> addMarkers(Set<_i4.Marker>? markersToAdd) =>
+  _i6.Future<void> addMarkers(Set<_i3.Marker>? markersToAdd) =>
       (super.noSuchMethod(
         Invocation.method(
           #addMarkers,
           [markersToAdd],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
+
   @override
-  _i5.Future<void> changeMarkers(Set<_i4.Marker>? markersToChange) =>
+  _i6.Future<void> changeMarkers(Set<_i3.Marker>? markersToChange) =>
       (super.noSuchMethod(
         Invocation.method(
           #changeMarkers,
           [markersToChange],
         ),
-        returnValue: _i5.Future<void>.value(),
-        returnValueForMissingStub: _i5.Future<void>.value(),
-      ) as _i5.Future<void>);
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
+
   @override
-  void removeMarkers(Set<_i4.MarkerId>? markerIdsToRemove) =>
+  void removeMarkers(Set<_i3.MarkerId>? markerIdsToRemove) =>
       super.noSuchMethod(
         Invocation.method(
           #removeMarkers,
@@ -368,24 +367,27 @@ class MockMarkersController extends _i1.Mock implements _i3.MarkersController {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void showMarkerInfoWindow(_i4.MarkerId? markerId) => super.noSuchMethod(
+  void showMarkerInfoWindow(_i3.MarkerId? markerId) => super.noSuchMethod(
         Invocation.method(
           #showMarkerInfoWindow,
           [markerId],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void hideMarkerInfoWindow(_i4.MarkerId? markerId) => super.noSuchMethod(
+  void hideMarkerInfoWindow(_i3.MarkerId? markerId) => super.noSuchMethod(
         Invocation.method(
           #hideMarkerInfoWindow,
           [markerId],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  bool isInfoWindowShown(_i4.MarkerId? markerId) => (super.noSuchMethod(
+  bool isInfoWindowShown(_i3.MarkerId? markerId) => (super.noSuchMethod(
         Invocation.method(
           #isInfoWindowShown,
           [markerId],
@@ -393,10 +395,11 @@ class MockMarkersController extends _i1.Mock implements _i3.MarkersController {
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
+
   @override
   void bindToMap(
     int? mapId,
-    _i2.GMap? googleMap,
+    _i4.Map? googleMap,
   ) =>
       super.noSuchMethod(
         Invocation.method(
@@ -414,33 +417,30 @@ class MockMarkersController extends _i1.Mock implements _i3.MarkersController {
 ///
 /// See the documentation for Mockito's code generation for more information.
 class MockTileOverlaysController extends _i1.Mock
-    implements _i3.TileOverlaysController {
+    implements _i2.TileOverlaysController {
   @override
-  _i2.GMap get googleMap => (super.noSuchMethod(
+  _i4.Map get googleMap => (super.noSuchMethod(
         Invocation.getter(#googleMap),
-        returnValue: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-        returnValueForMissingStub: _FakeGMap_0(
-          this,
-          Invocation.getter(#googleMap),
-        ),
-      ) as _i2.GMap);
+        returnValue: _i5.mapShim(),
+        returnValueForMissingStub: _i5.mapShim(),
+      ) as _i4.Map);
+
   @override
-  set googleMap(_i2.GMap? _googleMap) => super.noSuchMethod(
+  set googleMap(_i4.Map? _googleMap) => super.noSuchMethod(
         Invocation.setter(
           #googleMap,
           _googleMap,
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get mapId => (super.noSuchMethod(
         Invocation.getter(#mapId),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
+
   @override
   set mapId(int? _mapId) => super.noSuchMethod(
         Invocation.setter(
@@ -449,8 +449,9 @@ class MockTileOverlaysController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void addTileOverlays(Set<_i4.TileOverlay>? tileOverlaysToAdd) =>
+  void addTileOverlays(Set<_i3.TileOverlay>? tileOverlaysToAdd) =>
       super.noSuchMethod(
         Invocation.method(
           #addTileOverlays,
@@ -458,8 +459,9 @@ class MockTileOverlaysController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void changeTileOverlays(Set<_i4.TileOverlay>? tileOverlays) =>
+  void changeTileOverlays(Set<_i3.TileOverlay>? tileOverlays) =>
       super.noSuchMethod(
         Invocation.method(
           #changeTileOverlays,
@@ -467,8 +469,9 @@ class MockTileOverlaysController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void removeTileOverlays(Set<_i4.TileOverlayId>? tileOverlayIds) =>
+  void removeTileOverlays(Set<_i3.TileOverlayId>? tileOverlayIds) =>
       super.noSuchMethod(
         Invocation.method(
           #removeTileOverlays,
@@ -476,18 +479,20 @@ class MockTileOverlaysController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
-  void clearTileCache(_i4.TileOverlayId? tileOverlayId) => super.noSuchMethod(
+  void clearTileCache(_i3.TileOverlayId? tileOverlayId) => super.noSuchMethod(
         Invocation.method(
           #clearTileCache,
           [tileOverlayId],
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void bindToMap(
     int? mapId,
-    _i2.GMap? googleMap,
+    _i4.Map? googleMap,
   ) =>
       super.noSuchMethod(
         Invocation.method(
