@@ -3,7 +3,8 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:js_util' show getProperty;
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -191,8 +192,8 @@ void main() {
         final gmaps.MapTypeStyle style = styles[0];
         expect(style.featureType, 'poi.park');
         expect(style.elementType, 'labels.text.fill');
-        expect(style.stylers?.length, 1);
-        expect(getProperty<String>(style.stylers![0]!, 'color'), '#6b9a76');
+        expect(style.stylers.length, 1);
+        expect((style.stylers[0]['color']! as JSString).toDart, '#6b9a76');
       });
 
       testWidgets('throws MapStyleException for invalid styles',
