@@ -95,37 +95,42 @@ class MockGoogleMapController extends _i1.Mock
           Invocation.getter(#configuration),
         ),
       ) as _i2.MapConfiguration);
+
   @override
-  _i3.StreamController<_i2.MapEvent<dynamic>> get stream => (super.noSuchMethod(
+  _i3.StreamController<_i2.MapEvent<Object?>> get stream => (super.noSuchMethod(
         Invocation.getter(#stream),
-        returnValue: _FakeStreamController_1<_i2.MapEvent<dynamic>>(
+        returnValue: _FakeStreamController_1<_i2.MapEvent<Object?>>(
           this,
           Invocation.getter(#stream),
         ),
         returnValueForMissingStub:
-            _FakeStreamController_1<_i2.MapEvent<dynamic>>(
+            _FakeStreamController_1<_i2.MapEvent<Object?>>(
           this,
           Invocation.getter(#stream),
         ),
-      ) as _i3.StreamController<_i2.MapEvent<dynamic>>);
+      ) as _i3.StreamController<_i2.MapEvent<Object?>>);
+
   @override
-  _i3.Stream<_i2.MapEvent<dynamic>> get events => (super.noSuchMethod(
+  _i3.Stream<_i2.MapEvent<Object?>> get events => (super.noSuchMethod(
         Invocation.getter(#events),
-        returnValue: _i3.Stream<_i2.MapEvent<dynamic>>.empty(),
-        returnValueForMissingStub: _i3.Stream<_i2.MapEvent<dynamic>>.empty(),
-      ) as _i3.Stream<_i2.MapEvent<dynamic>>);
+        returnValue: _i3.Stream<_i2.MapEvent<Object?>>.empty(),
+        returnValueForMissingStub: _i3.Stream<_i2.MapEvent<Object?>>.empty(),
+      ) as _i3.Stream<_i2.MapEvent<Object?>>);
+
   @override
   bool get isInitialized => (super.noSuchMethod(
         Invocation.getter(#isInitialized),
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
+
   @override
   List<_i5.MapTypeStyle> get styles => (super.noSuchMethod(
         Invocation.getter(#styles),
         returnValue: <_i5.MapTypeStyle>[],
         returnValueForMissingStub: <_i5.MapTypeStyle>[],
       ) as List<_i5.MapTypeStyle>);
+
   @override
   void debugSetOverrides({
     _i4.DebugCreateMapFunction? createMap,
@@ -154,6 +159,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void init() => super.noSuchMethod(
         Invocation.method(
@@ -162,6 +168,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void updateMapConfiguration(_i2.MapConfiguration? update) =>
       super.noSuchMethod(
@@ -171,6 +178,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void updateStyles(List<_i5.MapTypeStyle>? styles) => super.noSuchMethod(
         Invocation.method(
@@ -179,6 +187,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.Future<_i2.LatLngBounds> getVisibleRegion() => (super.noSuchMethod(
         Invocation.method(
@@ -201,6 +210,7 @@ class MockGoogleMapController extends _i1.Mock
           ),
         )),
       ) as _i3.Future<_i2.LatLngBounds>);
+
   @override
   _i3.Future<_i2.ScreenCoordinate> getScreenCoordinate(_i2.LatLng? latLng) =>
       (super.noSuchMethod(
@@ -225,6 +235,7 @@ class MockGoogleMapController extends _i1.Mock
           ),
         )),
       ) as _i3.Future<_i2.ScreenCoordinate>);
+
   @override
   _i3.Future<_i2.LatLng> getLatLng(_i2.ScreenCoordinate? screenCoordinate) =>
       (super.noSuchMethod(
@@ -247,6 +258,7 @@ class MockGoogleMapController extends _i1.Mock
           ),
         )),
       ) as _i3.Future<_i2.LatLng>);
+
   @override
   _i3.Future<void> moveCamera(_i2.CameraUpdate? cameraUpdate) =>
       (super.noSuchMethod(
@@ -257,6 +269,7 @@ class MockGoogleMapController extends _i1.Mock
         returnValue: _i3.Future<void>.value(),
         returnValueForMissingStub: _i3.Future<void>.value(),
       ) as _i3.Future<void>);
+
   @override
   _i3.Future<double> getZoomLevel() => (super.noSuchMethod(
         Invocation.method(
@@ -266,6 +279,7 @@ class MockGoogleMapController extends _i1.Mock
         returnValue: _i3.Future<double>.value(0.0),
         returnValueForMissingStub: _i3.Future<double>.value(0.0),
       ) as _i3.Future<double>);
+
   @override
   void updateCircles(_i2.CircleUpdates? updates) => super.noSuchMethod(
         Invocation.method(
@@ -274,6 +288,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void updatePolygons(_i2.PolygonUpdates? updates) => super.noSuchMethod(
         Invocation.method(
@@ -282,6 +297,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void updatePolylines(_i2.PolylineUpdates? updates) => super.noSuchMethod(
         Invocation.method(
@@ -290,6 +306,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i3.Future<void> updateMarkers(_i2.MarkerUpdates? updates) =>
       (super.noSuchMethod(
@@ -300,6 +317,7 @@ class MockGoogleMapController extends _i1.Mock
         returnValue: _i3.Future<void>.value(),
         returnValueForMissingStub: _i3.Future<void>.value(),
       ) as _i3.Future<void>);
+
   @override
   void updateClusterManagers(_i2.ClusterManagerUpdates? updates) =>
       super.noSuchMethod(
@@ -309,6 +327,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void updateTileOverlays(Set<_i2.TileOverlay>? newOverlays) =>
       super.noSuchMethod(
@@ -318,6 +337,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clearTileCache(_i2.TileOverlayId? id) => super.noSuchMethod(
         Invocation.method(
@@ -326,6 +346,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void showInfoWindow(_i2.MarkerId? markerId) => super.noSuchMethod(
         Invocation.method(
@@ -334,6 +355,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void hideInfoWindow(_i2.MarkerId? markerId) => super.noSuchMethod(
         Invocation.method(
@@ -342,6 +364,7 @@ class MockGoogleMapController extends _i1.Mock
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool isInfoWindowShown(_i2.MarkerId? markerId) => (super.noSuchMethod(
         Invocation.method(
@@ -351,6 +374,7 @@ class MockGoogleMapController extends _i1.Mock
         returnValue: false,
         returnValueForMissingStub: false,
       ) as bool);
+
   @override
   void dispose() => super.noSuchMethod(
         Invocation.method(
