@@ -56,7 +56,11 @@ void main() {
       MarkerController(marker: marker, onTap: onTap);
 
       // Trigger a click event...
-      gmaps.Event.trigger(marker, 'click', <Object?>[gmaps.MapMouseEvent()]);
+      gmaps.event.trigger(
+        marker,
+        'click',
+        gmaps.MapMouseEvent(),
+      );
 
       // The event handling is now truly async. Wait for it...
       expect(await methodCalled, isTrue);
@@ -66,8 +70,11 @@ void main() {
       MarkerController(marker: marker, onDragStart: onDragStart);
 
       // Trigger a drag end event...
-      gmaps.Event.trigger(marker, 'dragstart',
-          <Object?>[gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0)]);
+      gmaps.event.trigger(
+        marker,
+        'dragstart',
+        gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0),
+      );
 
       expect(await methodCalled, isTrue);
     });
@@ -76,10 +83,10 @@ void main() {
       MarkerController(marker: marker, onDrag: onDrag);
 
       // Trigger a drag end event...
-      gmaps.Event.trigger(
+      gmaps.event.trigger(
         marker,
         'drag',
-        <Object?>[gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0)],
+        gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0),
       );
 
       expect(await methodCalled, isTrue);
@@ -89,10 +96,10 @@ void main() {
       MarkerController(marker: marker, onDragEnd: onDragEnd);
 
       // Trigger a drag end event...
-      gmaps.Event.trigger(
+      gmaps.event.trigger(
         marker,
         'dragend',
-        <Object?>[gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0)],
+        gmaps.MapMouseEvent()..latLng = gmaps.LatLng(0, 0),
       );
 
       expect(await methodCalled, isTrue);
@@ -104,7 +111,7 @@ void main() {
         ..draggable = true
         ..position = gmaps.LatLng(42, 54);
 
-      expect(marker.draggable, isNull);
+      expect(marker.isDraggableDefined(), isFalse);
 
       controller.update(options);
 
@@ -124,7 +131,7 @@ void main() {
 
     testWidgets('showInfoWindow', (WidgetTester tester) async {
       final gmaps.InfoWindow infoWindow = gmaps.InfoWindow();
-      final gmaps.GMap map = gmaps.GMap(createDivElement());
+      final gmaps.Map map = gmaps.Map(createDivElement());
       marker.set('map', map);
       final MarkerController controller = MarkerController(
         marker: marker,
@@ -139,7 +146,7 @@ void main() {
 
     testWidgets('hideInfoWindow', (WidgetTester tester) async {
       final gmaps.InfoWindow infoWindow = gmaps.InfoWindow();
-      final gmaps.GMap map = gmaps.GMap(createDivElement());
+      final gmaps.Map map = gmaps.Map(createDivElement());
       marker.set('map', map);
       final MarkerController controller = MarkerController(
         marker: marker,
@@ -157,7 +164,7 @@ void main() {
 
       setUp(() {
         final gmaps.InfoWindow infoWindow = gmaps.InfoWindow();
-        final gmaps.GMap map = gmaps.GMap(createDivElement());
+        final gmaps.Map map = gmaps.Map(createDivElement());
         marker.set('map', map);
         controller = MarkerController(marker: marker, infoWindow: infoWindow);
       });
