@@ -35,14 +35,18 @@ void main() {
     late gmaps.Circle circle;
 
     setUp(() {
-      circle = gmaps.Circle();
+      circle = gmaps.Circle(gmaps.CircleOptions());
     });
 
     testWidgets('onTap gets called', (WidgetTester tester) async {
       CircleController(circle: circle, consumeTapEvents: true, onTap: onTap);
 
       // Trigger a click event...
-      gmaps.Event.trigger(circle, 'click', <Object?>[gmaps.MapMouseEvent()]);
+      gmaps.event.trigger(
+        circle,
+        'click',
+        gmaps.MapMouseEvent(),
+      );
 
       // The event handling is now truly async. Wait for it...
       expect(await methodCalled, isTrue);
@@ -53,7 +57,7 @@ void main() {
       final gmaps.CircleOptions options = gmaps.CircleOptions()
         ..draggable = true;
 
-      expect(circle.draggable, isNull);
+      expect(circle.isDraggableDefined(), isFalse);
 
       controller.update(options);
 
@@ -98,7 +102,11 @@ void main() {
       PolygonController(polygon: polygon, consumeTapEvents: true, onTap: onTap);
 
       // Trigger a click event...
-      gmaps.Event.trigger(polygon, 'click', <Object?>[gmaps.MapMouseEvent()]);
+      gmaps.event.trigger(
+        polygon,
+        'click',
+        gmaps.MapMouseEvent(),
+      );
 
       // The event handling is now truly async. Wait for it...
       expect(await methodCalled, isTrue);
@@ -109,7 +117,7 @@ void main() {
       final gmaps.PolygonOptions options = gmaps.PolygonOptions()
         ..draggable = true;
 
-      expect(polygon.draggable, isNull);
+      expect(polygon.isDraggableDefined(), isFalse);
 
       controller.update(options);
 
@@ -158,7 +166,11 @@ void main() {
       );
 
       // Trigger a click event...
-      gmaps.Event.trigger(polyline, 'click', <Object?>[gmaps.MapMouseEvent()]);
+      gmaps.event.trigger(
+        polyline,
+        'click',
+        gmaps.MapMouseEvent(),
+      );
 
       // The event handling is now truly async. Wait for it...
       expect(await methodCalled, isTrue);
@@ -171,7 +183,7 @@ void main() {
       final gmaps.PolylineOptions options = gmaps.PolylineOptions()
         ..draggable = true;
 
-      expect(polyline.draggable, isNull);
+      expect(polyline.isDraggableDefined(), isFalse);
 
       controller.update(options);
 
